+++ sched/event/event_wait.c
@@ -33,7 +33,7 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: nxevent_tickwait
+ * Name: nxevent_tickwait_wait
  *
  * Description:
  *   Wait for all of the specified events for the specified tick time.
@@ -45,6 +45,7 @@
  *
  * Input Parameters:
  *   event  - Address of the event object
+ *   wait   - Address of the event wait object
  *   events - Set of events to wait
  *          - Set events to 0 will indicate wait from any events
  *   eflags - Events flags
@@ -59,15 +60,18 @@
  *
  ****************************************************************************/
 
-nxevent_mask_t nxevent_tickwait(FAR nxevent_t *event, nxevent_mask_t events,
-                                nxevent_flags_t eflags, uint32_t delay)
+nxevent_mask_t nxevent_tickwait_wait(FAR nxevent_t *event,
+                                     FAR nxevent_wait_t *wait,
+                                     nxevent_mask_t events,
+                                     nxevent_flags_t eflags,
+                                     uint32_t delay)
 {
-  nxevent_wait_t wait;
   irqstate_t flags;
   bool waitany;
   int ret;
 
-  DEBUGASSERT(event != NULL && up_interrupt_context() == false);
+  DEBUGASSERT(event != NULL && wait != NULL &&
+              up_interrupt_context() == false);
 
   waitany = ((eflags & NXEVENT_WAIT_ALL) == 0);
 
@@ -117,31 +121,31 @@ nxevent_mask_t nxevent_tickwait(FAR nxevent_t *event, nxevent_mask_t events,
     {
       /* Initialize event wait */
 
-      nxsem_init(&wait.sem, 0, 0);
-      wait.expect = events;
-      wait.eflags = eflags;
+      nxsem_init(&(wait->sem), 0, 0);
+      wait->expect = events;
+      wait->eflags = eflags;
 
-      list_add_tail(&event->list, &wait.node);
+      list_add_tail(&event->list, &(wait->node));
 
       /* Wait for the event */
 
       if (delay == UINT32_MAX)
         {
-          ret = nxsem_wait_uninterruptible(&wait.sem);
+          ret = nxsem_wait_uninterruptible(&(wait->sem));
         }
       else
         {
-          ret = nxsem_tickwait_uninterruptible(&wait.sem, delay);
+          ret = nxsem_tickwait_uninterruptible(&(wait->sem), delay);
         }
 
       /* Destroy local variables */
 
-      nxsem_destroy(&wait.sem);
-      list_delete(&wait.node);
+      nxsem_destroy(&(wait->sem));
+      list_delete(&(wait->node));
 
       if (ret == 0)
         {
-          events = wait.expect;
+          events = wait->expect;
         }
       else
         {
@@ -152,6 +156,41 @@ nxevent_mask_t nxevent_tickwait(FAR nxevent_t *event, nxevent_mask_t events,
   leave_critical_section(flags);
 
   return events;
+}
+
+/****************************************************************************
+ * Name: nxevent_tickwait
+ *
+ * Description:
+ *   Wait for all of the specified events for the specified tick time.
+ *
+ *   This routine waits on event object event until all of the specified
+ *   events have been delivered to the event object, or the maximum wait time
+ *   timeout has expired. A thread may wait on up to 32 distinctly numbered
+ *   events that are expressed as bits in a single 32-bit word.
+ *
+ * Input Parameters:
+ *   event  - Address of the event object
+ *   events - Set of events to wait
+ *          - Set events to 0 will indicate wait from any events
+ *   eflags - Events flags
+ *   delay  - Ticks to wait from the start time until the event is
+ *            posted.  If ticks is zero, then this function is equivalent
+ *            to nxevent_trywait().
+ *
+ * Returned Value:
+ *   This is an internal OS interface and should not be used by applications.
+ *   Return of matching events upon success.
+ *   0 if matching events were not received within the specified time.
+ *
+ ****************************************************************************/
+
+nxevent_mask_t nxevent_tickwait(FAR nxevent_t *event, nxevent_mask_t events,
+                                nxevent_flags_t eflags, uint32_t delay)
+{
+  nxevent_wait_t wait;
+
+  return nxevent_tickwait_wait(event, &wait, events, eflags, delay);
 }
 
 /****************************************************************************
